change take_positional_word to do a scan
instead of picking the next unused one.

`positional` should be used either after all the named items or as an
item adjacent to a named item in some way, still after all the named
items within the scope.

In case of valid user input scanning for or picking up the front item
gives the same results. When user specifies an invalid item - it will
not be consumed during normal scanning so picking the front item fails.

Error from positional item failing with "not present" will take the
priority over "not expected" from invalid item resulting in a bad error
message.